requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-experimental>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
